Rename file Plugin to InstalledPlugin.
(cherry picked from commit 0229354da0b125357242a14696306007ec558bd9)
