+++ plugins/woocommerce-admin/client/marketing/types/InstalledPlugin.ts
