+++ plugins/woocommerce-admin/client/marketing/types/index.ts
@@ -1,1 +1,1 @@
-export { InstalledPlugin } from './Plugin';
+export { InstalledPlugin } from './InstalledPlugin';
